+++ demos/backend/src/slow.ts
@@ -1,5 +1,5 @@
-import { Logger } from '../../../packages/logger/src'
-import { RocksDB } from '../../../packages/rocksdb/src'
+import { Logger } from '../../../packages/extension-logger/src'
+import { RocksDB } from '../../../packages/extension-rocksdb/src'
 import { Server } from '../../../packages/server/src'
 
 const server = Server.configure({